requirements.txt
mysql-connector-python
numpy
pandas
dotenv
tkinter